Fixed compilation of MFC-based modules (MFCdemo application and NicoletOne source module).
git-svn-id: http://www.bci2000.org/svn/trunk@3232 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/contrib/Application/MFCdemo/MFCdemoDlg.cpp
@@ -80,15 +80,15 @@ void CMFCdemoDlg::OnPaint()
 	}
 	else
 	{
-    // Draw the BCI cursor icon.
-    CPaintDC dc( this );
+		// Draw the BCI cursor icon.
+		CPaintDC dc( this );
 		int cxIcon = GetSystemMetrics(SM_CXICON);
 		int cyIcon = GetSystemMetrics(SM_CYICON);
-    CRect rect;
-    GetClientRect( &rect );
-    int xPos = rect.left + rect.Width() * mCursorX - cxIcon / 2,
-        yPos = rect.top + rect.Height() * mCursorY - cyIcon / 2;
-    dc.DrawIcon( xPos, yPos, mCursorIcon );
+		CRect rect;
+		GetClientRect( &rect );
+		int xPos = static_cast<int>( rect.left + rect.Width() * mCursorX - cxIcon / 2 ),
+			yPos = static_cast<int>( rect.top + rect.Height() * mCursorY - cyIcon / 2 );
+		dc.DrawIcon( xPos, yPos, mCursorIcon );
 		CDialog::OnPaint();
 	}
 }
